feat: Add local model inference using HuggingFace API
The commit introduce the local model script in the model that allows you to keep using the local models in the API without the need to call the HF inference or any other commercial API model. This will help save costs. Also the transformer is made an optional dependency for the project.